ALOT of freehand changes, none of which are tested and im sure will break.
new set-listerner built to accept args from the pieepline. and heaps added tot he new listerner including creating a socket listerner and adding it to a global array

we need to test it, break it and fix it before moving on.

the crux of this is to have the raw socket listernerr working.

after that we need to buld the method for upgrading the tcp reverse connection to a tty agent.
this will also be the time we begin creating the New-Agent payload.

after that we can do a get-aggent commandlet too.

remember that rawtcp connections are stored in the the listerner. but upgraded connections will be in the Get-Agent commandlet